Some code re-arrangments
Moved around a few things after running with some new Debug flags (which might
be a tad excessive) and receiving some errors.

Also moved all equality checks around to have the variable to check on the
right (where possible), which will be the default from now on so as not to
accidentally use a single '=' on accident.

+++ includes/Memory.h
@@ -6,7 +6,7 @@
 #include "Structs.h"
 #include "Enums.h"
 
-extern uint16_t selected;
+extern int selected;
 extern uint16_t output_height;
 extern uint16_t *memory_output;
 extern uint16_t selected_address;
